+++ R/mentions.R
@@ -25,7 +25,7 @@ get_mentions <- function(n = 200,
                          since_id = NULL,
                          max_id = NULL,
                          parse = TRUE,
-                         retryonratelimit = FALSE,
+                         retryonratelimit = NULL,
                          verbose = TRUE,
                          token = NULL,
                          ...) {